cloud-function/requirements.txt
google-api-python-client==2.96.0
google-auth==2.23.4
google-auth-httplib2==0.1.0
